Rename soci-cli to soci
