+++ cmd/soci/commands/image/image.go
