+++ cmd/soci/commands/image/list_indices.go
