+++ cmd/soci/commands/index/index.go
