+++ cmd/soci/commands/index/list.go
