+++ cmd/soci/commands/ztoc/info.go
