+++ cmd/soci/commands/ztoc/ztoc.go
